Token missing and null checking fix

+++ src/cypress/commands/deleteDocumentTypesByNamePrefix.ts
@@ -6,21 +6,24 @@ export default class DeleteDocumentTypesByNamePrefix extends CommandBase {
 
   method(prefix) {
     const cy = this.cy;
-
-    return cy
-      .request({
-        method: 'GET',
-        url:
-          this.relativeBackOfficePath +
-          '/backoffice/UmbracoTrees/ContentTypeTree/GetNodes?id=-1&application=settings&tree=&use=main&culture=',
-      })
-      .then((response) => {
-        const items = JsonHelper.getBody(response);
-        for (const item of items) {
-          if (item.name.startsWith(prefix)) {
-            cy.deleteDocumentTypeById(item.id);
+    return cy.getCookie('UMB-XSRF-TOKEN', { log: false }).then((token) => {
+      return cy
+        .request({
+          method: 'GET',
+          url:
+            this.relativeBackOfficePath +
+            '/backoffice/UmbracoTrees/ContentTypeTree/GetNodes?id=-1&application=settings&tree=&use=main&culture=',
+          headers: {
+            Accept: 'application/json',
+            'X-UMB-XSRF-TOKEN': token.value,
+          },
+        })
+        .then((response) => {
+          const items = JsonHelper.getBody(response);
+          for (const item of items) {
+            if (item.name?.startsWith(prefix)) cy.deleteDocumentTypeById(item.id);
           }
-        }
-      });
+        });
+    });
   }
 }
